refactor: decompose lex.go into object.go
Extract PDF object type definitions and parse methods into
object.go: object, dict, array, stream, objptr, objdef types,
maxObjectDepth, readObject, readKeywordObject, tryReadObjRef,
readArray, readDict.

lex.go retains the tokenizer only: buffer, token/name/keyword,
and all byte/token-level read methods.
